move automate radio button to miq.wt
